refactor(frameworks): rename components to frameworks
Makes more sense as the folder contains the components of frameworks
