+++ src/frameworks/bootstrap/checkboxes.html
